rv-predict-engine: Moved engines and main files out of the default package
