+++ rv-predict-engine/src/engines/CPRaceDetect.java
@@ -1,4 +1,4 @@
-/*******************************************************************************
+package engines; /*******************************************************************************
  * Copyright (c) 2013 University of Illinois
  * 
  * All rights reserved.
@@ -31,36 +31,27 @@ import java.io.PrintWriter;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.Set;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.Vector;
-import java.util.Map.Entry;
 
 import config.Configuration;
-import trace.AbstractNode;
-import trace.IMemNode;
-import trace.LockPair;
 import trace.ReadNode;
 import trace.Trace;
 import trace.TraceInfo;
 import trace.WriteNode;
-import trace.AbstractNode.TYPE;
-import violation.AtomicityViolation;
-import violation.Deadlock;
 import violation.ExactRace;
 import violation.IViolation;
 import violation.Race;
-import z3.Z3Engine;
 import db.DBEngine;
 
 /**
- * HBRaceDetect implements the happens-before methods for race detection.
+ * engines.CPRaceDetect implements the causal-precedes methods for race detection.
  * 
  * @author jeffhuang
  *
  */
-public class HBRaceDetect {
+public class CPRaceDetect {
 
 	private static Configuration config;
 	private static HashSet<IViolation> races= new HashSet<IViolation>();
@@ -70,7 +61,7 @@ public class HBRaceDetect {
 		try{
 		String fname = "result."+(config.window_size/1000)+"k";
 		out = new PrintWriter(new FileWriter(fname,true));
-		out.println("\n------------------ HB: "+appname+" -------------------\n");
+		out.println("\n------------------ CP: "+appname+" -------------------\n");
 		}catch(Exception e)
 		{
 			e.printStackTrace();
@@ -96,13 +87,13 @@ public class HBRaceDetect {
 			return schedule;
 	}
 	/**
-	 * traverse all conflicting pairs. For each pair, query the CPEngine 
+	 * traverse all conflicting pairs. For each pair, query the CPEngine
 	 * whether there are reachable or not. If yes, report a race.
 	 * 
 	 * @param engine
 	 * @param trace
 	 */
-	private static void detectRace(HBEngine engine, Trace trace)
+	private static void detectRace(CPEngine engine, Trace trace)
 	{
 		Iterator<String> 
 		addrIt =trace.getIndexedThreadReadWriteNodes().keySet().iterator();
@@ -117,6 +108,7 @@ public class HBRaceDetect {
 			//continue if volatile
 			if(dotPos>-1&&trace.isAddressVolatile(addr.substring(dotPos+1))) continue;
 			}
+			
 			//get all read nodes on the address
 			Vector<ReadNode> readnodes = trace.getIndexedReadNodes().get(addr);
 			
@@ -158,7 +150,8 @@ public class HBRaceDetect {
 									races.add(race2); 
 
 								}
-								else races.add(race);
+								else
+								races.add(race);
 							
 							}
 						
@@ -191,7 +184,7 @@ public class HBRaceDetect {
 
 									}
 									else
-									races.add(race); 
+										races.add(race); 
 								}
 							}
 						}
@@ -261,7 +254,7 @@ public class HBRaceDetect {
 				
 				Trace trace = db.getTrace(index_start,index_end,info);
 					
-				HBEngine engine = new HBEngine(trace);				
+				CPEngine engine = new CPEngine(trace);				
 					
 				detectRace(engine,trace);
