+++ rv-predict-engine/src/rvpredict/engine/main/HBEngine.java
@@ -1,4 +1,4 @@
-/*******************************************************************************
+package rvpredict.engine.main; /*******************************************************************************
  * Copyright (c) 2013 University of Illinois
  * 
  * All rights reserved.
